requirements.txt
fastapi
uvicorn
numpy
torch
transformers
python-dotenv
langchain_openai
sentence_transformers
